requirements.txt
apache-beam[gcp]==2.55.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.20.0
google-cloud-pubsub==2.21.0
google-cloud-functions==1.16.0
shapely==2.0.3
pyproj==3.6.1
fiona==1.9.6
geopandas==0.14.4
pandas==2.2.1
reportlab==4.1.0
requests==2.31.0
dbt-bigquery==1.7.0
